chore!: move `PartialWitnessGenerator` code into `pwg.rs`
